refactor: consolidate FLUX.1/FLUX.2 into single flux/ package with shared modules
- Merge flux2/ into flux/ — eliminate separate package, reuse shared code
- Extract FluxJointAttention into modules/attention.py for FLUX.1/FLUX.2
- Add AdaLayerNorm abstraction (_AdaLayerNormBase + thin subclasses)
- Replace custom FeedForward with TRT-LLM MLP via _make_ffn() factory
- Use diffusers Timesteps/TimestepEmbedding instead of custom embeddings
- Rename attention_flux.py → pos_embed_flux.py (file contains RoPE, not attention)
- Fix OOM in test_attention_backend_comparison (sequential pipeline loading)
- Add E2E pipeline tests comparing TRT-LLM vs HuggingFace reference (PSNR)
- All 25 tests pass, PSNR unchanged (FLUX.1=24.13dB, FLUX.2=15.69dB)

+++ tensorrt_llm/_torch/visual_gen/models/flux/pipeline_flux.py
@@ -6,6 +6,7 @@
 import time
 from typing import Optional, Tuple
 
+import numpy as np
 import torch
 from diffusers import AutoencoderKL, FlowMatchEulerDiscreteScheduler
 from diffusers.utils.torch_utils import randn_tensor
@@ -126,7 +127,7 @@ class FluxPipeline(BasePipeline):
         if PipelineComponent.VAE not in skip_components:
             logger.info("Loading VAE...")
             self.vae = AutoencoderKL.from_pretrained(
-                checkpoint_dir, subfolder=PipelineComponent.VAE, torch_dtype=torch.float32
+                checkpoint_dir, subfolder=PipelineComponent.VAE, torch_dtype=torch.bfloat16
             ).to(device)
 
             self.vae_scale_factor = 2 ** (len(self.vae.config.block_out_channels) - 1)
@@ -236,8 +237,6 @@ class FluxPipeline(BasePipeline):
         logger.info(f"Latents shape: {latents.shape}")
 
         # Prepare timesteps with dynamic shifting (FLUX uses mu parameter)
-        import numpy as np
-
         image_seq_len = latents.shape[1]
         mu = self._compute_mu(image_seq_len, num_inference_steps)
